(WIP) refac: updating notebook
