+++ notebooks/utils/__init__.py
@@ -1,0 +1,17 @@
+from .plot_utils import (
+    add_pixel_kde,
+    add_subplot_axes,
+    add_text,
+    boundary_color,
+    clean_ax,
+    get_zoomin_hw,
+)
+
+__all__ = [
+    "add_pixel_kde",
+    "clean_ax",
+    "add_subplot_axes",
+    "add_text",
+    "boundary_color",
+    "get_zoomin_hw",
+]
